api: fix `upload` to handle attributes `FilePath`, `FileName`
When we receive headers, they are automatically formatted from any case to
only the first letter being uppercase. For example, `FilePath` transforms into
`Filepath`. However, NEOFS is case-sensitive for these attributes, which is why
we intentionally change them to CamelCase style. We hope this is a temporary
workaround and will be removed asap.
Connected to https://github.com/nspcc-dev/neofs-http-gw/issues/255.

Signed-off-by: Tatiana Nesterenko <tatiana@nspcc.io>