api: Remove the `fullBearer` parameter from all API requests
This parameter indicates that a binary bearer token was requested via the
`/auth/bearer` request and is being used instead of wallet connect. However,
it's not difficult to detect this without the `fullBearer` parameter. If the
`X-Bearer-Signature` and `X-Bearer-Signature-Key` are absent in the request
header, we can assume that the received token is `fullBearer`.

Signed-off-by: Tatiana Nesterenko <tatiana@nspcc.io>