go.mod: Upgrade to almost latest `github.com/nspcc-dev/neofs-sdk-go`
Continues 662fc997f70b3b99f6742b3402665212bca07d58. Brings many fixes
and improvements. The main thing is the rejection of
`github.com/nspcc-dev/neofs-api-go/v2` module, it's not even in indirect
deps anymore. It was necessary to fix a small section of code that was
previously based on a type from this module.

Update to the latest TBD after #235 because subsequent NeoFS SDK
revisions require Go 1.23.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>